Move `MultiVMBaseSystemContracts` to executor
