+++ core/lib/vm_executor/src/oneshot/mod.rs
@@ -32,8 +32,9 @@ use zksync_types::{
 };
 use zksync_utils::{h256_to_u256, u256_to_h256};
 
-pub use self::mock::MockOneshotExecutor;
+pub use self::{contracts::MultiVMBaseSystemContracts, mock::MockOneshotExecutor};
 
+mod contracts;
 mod metrics;
 mod mock;
